requirements.txt
python-telegram-bot==12.8
redis==3.5.3
requests~=2.24.0
python-dotenv~=0.13.0